+++ Tests/Runtime/GameSimManagerTests.cs
@@ -7,6 +7,7 @@ namespace Unity.Simulation.Games.Tests
 {
     public class GameSimManagerTests
     {
+#if UNITY_GAME_SIMULATION || UNITY_EDITOR
         [UnityTest]
         public IEnumerator VerifyResetAndFlush_ResetsCounter_AndSavesCurrentStateTo_FileSystem()
         {
@@ -15,5 +16,6 @@ namespace Unity.Simulation.Games.Tests
             yield return null;
             Assert.IsTrue(GameSimManager.Instance.GetCounter("test").Value == 0);
         }
+#endif
     }
 }