+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.3-preview.2] - 2020-07-15
+## This is the v0.4.3-preview.2 release of *Unity Game Simulation*
+- Uses a separate remote config environment, allowing users to use remote config without polluting their game environment from their game sim environment
+- Remote Config package requirement updated to 1.3.2-preview.6
+- Builds will no longer attempt to upload on failure
+- Add additional error messages
+- Add scripting define: UNITY_GAME_SIMULATION -- if this define is not present, all public API are no-op's. This is handled by the Game Sim Editor Window's "Build and Upload" button. See https://docs.unity3d.com/ScriptReference/PlayerSettings.SetScriptingDefineSymbolsForGroup.html for integration into CI/CD.
+
 ## [0.4.3-preview.1] - 2020-06-25
 
 ## This is the v0.4.3-preview.1 release of *Unity Game Simulation*